GreenAccounter-backend-connection/requirements.txt
pandas
numpy
psutil
argparse
firebase-admin
asyncio
flask
flask-cors
aioelectricitymaps
paramiko
python-dotenv